[CALCITE-1943] Add back NavigationExpander and NavigationReplacer in SqlValidatorImpl (Dian Fu)
It will replace clauses such as A.price with PREV(A.price, 0) and
makes the implementation of RexVisitor.visitPatternFieldRef(RexPatternFieldRef)
more unified.  Otherwise, it is difficult to implement this method. If
it returns the specified field, then the navigation such as
PREV(A.price, 1) becomes impossible; if not, then comparisons such as
A.price > PREV(A.price, 1) becomes meaningless.

Close apache/calcite#516